feat: initial commit working project

+++ Services/MarketDataHelper.cs
@@ -12,7 +12,7 @@ namespace ZerodhaOxySocket
                 var kite = new Kite(apiKey, accessToken);
                 var quotes = kite.GetQuote(new string[] { $"{exchange}:{tradingsymbol}" });
                 if (quotes != null && quotes.TryGetValue($"{exchange}:{tradingsymbol}", out var q))
-                    return q.LastPrice;
+                    return (double)q.LastPrice;
             }
             catch (Exception ex) { Console.WriteLine("GetSpotPrice failed: " + ex.Message); }
             return 0;
